Rowing Quiz Added

+++ rowing_quiz/models.py
@@ -4,7 +4,7 @@ from django.db import models
 # Create your models here.
 class Crew(models.Model):
     crew_name = models.CharField(max_length=200)
-    score = models.IntegerField(default=0)
+    score = models.FloatField(default=0.0)
 
     def __str__(self):
         return self.crew_name
